Moved rootmod installer out of index.ts to clean it up.
General cleanup.
